pyproject.toml
[tool.poetry]
name = "boiler-bot"
version = "0.1.0"
description = "A shitty discord bot for my discord server"
authors = ["Samrid Pandit <samrid.pandit@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
jishaku = "^2.2.0"
pydantic = "^1.8.2"
cachetools = "^4.2.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.6"}
wavelink = "^0.9.10"
python-dotenv = "^0.19.0"
aerich = "^0.5.5"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}
isort = "^5.9.3"
pylint = "^2.9.6"
watchgod = "^0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
